test(ui): split native shell host coverage
