+++ ui/src/app/app-host.test.ts
@@ -1,7 +1,6 @@
 /* @vitest-environment jsdom */
 
 import type { RouteLocation, RouterState } from "@openclaw/uirouter";
-import { render } from "lit";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import type { GatewayBrowserClient } from "../api/gateway.ts";
 import type { AgentsListResult, GatewayAgentRow } from "../api/types.ts";
@@ -16,6 +15,7 @@ import {
   UI_COMMAND_EVENT,
 } from "../components/panel-toggle-contract.ts";
 import { createStorageMock } from "../test-helpers/storage.ts";
+import { resetAppHostTestGlobals, type ShellKeyboardState } from "./app-host.test-support.ts";
 import { selectShellRouteState } from "./app-host.ts";
 import type {
   ApplicationContext,
@@ -23,7 +23,6 @@ import type {
   ApplicationGatewaySnapshot,
 } from "./context.ts";
 import { shouldMergeChatChrome } from "./mobile-nav-layout.ts";
-import { navigationSurfaceIsHidden, renderFloatingUpdateCard } from "./navigation-surface.ts";
 import { resolveOnboardingMode } from "./onboarding-mode.ts";
 import { resetServerUiPrefsSync } from "./server-prefs.ts";
 
@@ -51,13 +50,6 @@ type ShellInitializationState = {
 type ShellServerPreferencesState = {
   runtime: { context: ApplicationContext };
   reconcileServerUiPrefs: (runtimeConfig: ApplicationContext["runtimeConfig"]) => void;
-};
-
-type ShellKeyboardState = {
-  runtime: {
-    context: ApplicationContext;
-  };
-  handleDocumentKeydown: (event: KeyboardEvent) => void;
 };
 
 type TestOptionalCustomElement = {
@@ -151,20 +143,6 @@ function createLazyElementSpec(label: string): TestOptionalCustomElement {
   };
 }
 
-type ShellNavigationState = {
-  runtime: {
-    context: ApplicationContext;
-  };
-  handleNativeToggleSidebar: () => void;
-  handleNativeOpenSearch: () => void;
-  handleNativeToggleSearch: (event: Event) => void;
-  handleNativeNewSession: () => void;
-  handleNativeHistoryState: (event: Event) => void;
-  nativeHistoryState: { canGoBack: boolean; canGoForward: boolean };
-  onboarding: boolean;
-  updated: () => void;
-};
-
 type ShellChromeEventState = {
   runtime: { context: ApplicationContext };
   navDrawerOpen: boolean;
@@ -188,31 +166,8 @@ type ShellSettingsSearchLoadState = {
   handleSettingsSearchQueryChange: (query: string) => Promise<void>;
 };
 
-type TestWebKitWindow = Window & {
-  webkit?: {
-    messageHandlers: {
-      openclawNav: { postMessage: (message: unknown) => void };
-    };
-  };
-};
-
-type MacosTitlebarControlsState = HTMLElement & {
-  navCollapsed: boolean;
-  historyOnly: boolean;
-  onOpenPalette?: () => void;
-  onOpenNewSession?: () => void;
-  updateComplete: Promise<boolean>;
-};
-
 afterEach(() => {
-  vi.useRealTimers();
-  Reflect.deleteProperty(window, "webkit");
-  document.documentElement.classList.remove(
-    "openclaw-native-macos",
-    "openclaw-native-nav",
-    "openclaw-native-web-chrome",
-  );
-  vi.unstubAllGlobals();
+  resetAppHostTestGlobals();
 });
 
 type ShellEpochState = {
@@ -886,257 +841,5 @@ describe("OpenClaw shell keyboard shortcuts", () => {
     await vi.advanceTimersByTimeAsync(1);
 
     expect(harness.refreshList).toHaveBeenCalledOnce();
-  });
-
-  it("opens Settings with Shift-Command-Comma", () => {
-    const navigate = vi.fn();
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellKeyboardState;
-    shell.runtime = {
-      context: {
-        navigate,
-      } as unknown as ApplicationContext,
-    };
-    const event = new KeyboardEvent("keydown", {
-      key: "<",
-      code: "Comma",
-      metaKey: true,
-      shiftKey: true,
-      cancelable: true,
-    });
-
-    shell.handleDocumentKeydown(event);
-
-    expect(event.defaultPrevented).toBe(true);
-    expect(navigate).toHaveBeenCalledWith("config", undefined);
-  });
-
-  it("opens Settings with Ctrl-Shift-Comma", () => {
-    const navigate = vi.fn();
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellKeyboardState;
-    shell.runtime = {
-      context: {
-        navigate,
-      } as unknown as ApplicationContext,
-    };
-    const event = new KeyboardEvent("keydown", {
-      key: "<",
-      code: "Comma",
-      ctrlKey: true,
-      shiftKey: true,
-      cancelable: true,
-    });
-
-    shell.handleDocumentKeydown(event);
-
-    expect(event.defaultPrevented).toBe(true);
-    expect(navigate).toHaveBeenCalledWith("config", undefined);
-  });
-
-  it("toggles the navigation sidebar when the native macOS titlebar button fires", () => {
-    const snapshot = { navCollapsed: false };
-    const update = vi.fn((next: { navCollapsed: boolean }) => {
-      snapshot.navCollapsed = next.navCollapsed;
-    });
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellNavigationState;
-    shell.runtime = {
-      context: {
-        navigation: { snapshot, update },
-      } as unknown as ApplicationContext,
-    };
-
-    shell.handleNativeToggleSidebar();
-    expect(update).toHaveBeenLastCalledWith({ navCollapsed: true });
-
-    shell.handleNativeToggleSidebar();
-    expect(update).toHaveBeenLastCalledWith({ navCollapsed: false });
-  });
-
-  it("opens search and starts a session from native titlebar events", () => {
-    const navigate = vi.fn();
-    const openPalette = vi.fn();
-    const togglePalette = vi.fn();
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellNavigationState;
-    Object.defineProperty(shell, "commandPalette", {
-      configurable: true,
-      value: { openPalette, togglePalette },
-    });
-    shell.runtime = {
-      context: {
-        navigate,
-        agentSelection: { state: { selectedId: "agent/a" } },
-      } as unknown as ApplicationContext,
-    };
-    shell.handleNativeOpenSearch();
-    const toggleEvent = new CustomEvent("openclaw:native-toggle-search", { cancelable: true });
-    shell.handleNativeToggleSearch(toggleEvent);
-    shell.handleNativeNewSession();
-
-    expect(openPalette).toHaveBeenCalledOnce();
-    expect(togglePalette).toHaveBeenCalledOnce();
-    // preventDefault is the handled signal for the native legacy fallback.
-    expect(toggleEvent.defaultPrevented).toBe(true);
-    expect(navigate).toHaveBeenCalledWith("new-session", { search: "?agent=agent%2Fa" });
-  });
-
-  it("keeps the new-thread control in the native titlebar only while collapsed", async () => {
-    const onOpenPalette = vi.fn();
-    const onOpenNewSession = vi.fn();
-    const controls = document.createElement(
-      "openclaw-macos-titlebar-controls",
-    ) as unknown as MacosTitlebarControlsState;
-    controls.navCollapsed = false;
-    controls.historyOnly = false;
-    controls.onOpenPalette = onOpenPalette;
-    controls.onOpenNewSession = onOpenNewSession;
-    document.body.append(controls);
-    await controls.updateComplete;
-
-    controls.querySelector<HTMLButtonElement>(".macos-titlebar-controls__search")?.click();
-    expect(controls.querySelector(".macos-titlebar-controls__new-session")).toBeNull();
-
-    controls.navCollapsed = true;
-    await controls.updateComplete;
-    controls.querySelector<HTMLButtonElement>(".macos-titlebar-controls__new-session")?.click();
-
-    expect(onOpenPalette).toHaveBeenCalledOnce();
-    expect(onOpenNewSession).toHaveBeenCalledOnce();
-    controls.remove();
-  });
-
-  it("retains a native new-session request until a context exists", () => {
-    const navigate = vi.fn();
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellNavigationState;
-
-    shell.handleNativeNewSession();
-
-    shell.runtime = {
-      context: {
-        navigate,
-        agentSelection: { state: { selectedId: "main" } },
-      } as unknown as ApplicationContext,
-    };
-    shell.handleNativeNewSession();
-
-    expect(navigate).toHaveBeenCalledExactlyOnceWith("new-session", { search: "?agent=main" });
-  });
-
-  it("does not start a native session during onboarding", () => {
-    const navigate = vi.fn();
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellNavigationState;
-    shell.runtime = {
-      context: {
-        navigate,
-        agentSelection: { state: { selectedId: "main" } },
-      } as unknown as ApplicationContext,
-    };
-    shell.onboarding = true;
-
-    shell.handleNativeNewSession();
-
-    expect(navigate).not.toHaveBeenCalled();
-  });
-
-  it("updates native history state from the host event", () => {
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellNavigationState;
-    shell.handleNativeHistoryState(
-      new CustomEvent("openclaw:native-history-state", {
-        detail: { canGoBack: true, canGoForward: false },
-      }),
-    );
-
-    expect(shell.nativeHistoryState).toEqual({ canGoBack: true, canGoForward: false });
-  });
-
-  it("deduplicates native nav state reports", () => {
-    const postMessage = vi.fn();
-    (window as TestWebKitWindow).webkit = {
-      messageHandlers: { openclawNav: { postMessage } },
-    };
-    const snapshot = { navCollapsed: false, navWidth: 280 };
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellNavigationState;
-    shell.runtime = {
-      context: {
-        navigation: { snapshot },
-      } as unknown as ApplicationContext,
-    };
-
-    shell.updated();
-    shell.updated();
-    snapshot.navCollapsed = true;
-    shell.updated();
-
-    expect(postMessage.mock.calls).toEqual([
-      [{ type: "nav-state", collapsed: false, width: 280 }],
-      [{ type: "nav-state", collapsed: true, width: 280 }],
-    ]);
-  });
-
-  it("leaves plain Command-Comma to the browser", () => {
-    const navigate = vi.fn();
-    const shell = document.createElement("openclaw-app-shell") as unknown as ShellKeyboardState;
-    shell.runtime = {
-      context: {
-        navigate,
-      } as unknown as ApplicationContext,
-    };
-    const event = new KeyboardEvent("keydown", {
-      key: ",",
-      code: "Comma",
-      metaKey: true,
-      cancelable: true,
-    });
-
-    shell.handleDocumentKeydown(event);
-
-    expect(event.defaultPrevented).toBe(false);
-    expect(navigate).not.toHaveBeenCalled();
-  });
-});
-
-describe("OpenClaw shell update affordance", () => {
-  it("renders a floating card only while desktop navigation is collapsed", () => {
-    const container = document.createElement("div");
-    const shared = {
-      onboarding: false,
-      updateAvailable: {
-        currentVersion: "2026.7.1",
-        latestVersion: "2026.7.2",
-        channel: "stable" as const,
-      },
-      updateRunning: false,
-      onUpdate: vi.fn(),
-    };
-    const collapsed = navigationSurfaceIsHidden({
-      navCollapsed: true,
-      navDrawerOpen: false,
-      mobileNavLayout: false,
-    });
-    render(renderFloatingUpdateCard({ ...shared, navigationSurfaceHidden: collapsed }), container);
-    expect(container.querySelector("openclaw-sidebar-update-card")).not.toBeNull();
-
-    const visible = navigationSurfaceIsHidden({
-      navCollapsed: false,
-      navDrawerOpen: false,
-      mobileNavLayout: false,
-    });
-    render(renderFloatingUpdateCard({ ...shared, navigationSurfaceHidden: visible }), container);
-    expect(container.querySelector("openclaw-sidebar-update-card")).toBeNull();
-  });
-
-  it("treats a closed mobile drawer as hidden navigation", () => {
-    expect(
-      navigationSurfaceIsHidden({
-        navCollapsed: false,
-        navDrawerOpen: false,
-        mobileNavLayout: true,
-      }),
-    ).toBe(true);
-    expect(
-      navigationSurfaceIsHidden({
-        navCollapsed: false,
-        navDrawerOpen: true,
-        mobileNavLayout: true,
-      }),
-    ).toBe(false);
   });
 });
